+++ src/Contracts/PermissionChecker.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Dlnsk\HierarchicalRBAC\Contracts;
+
+interface PermissionChecker
+{
+    public function check($ability, $arguments);
+}