+++ tests/TestCase.php
@@ -8,16 +8,6 @@ use Dlnsk\HierarchicalRBAC\HRBACServiceProvider;
 abstract class TestCase extends \Orchestra\Testbench\TestCase
 {
     /**
-     * Setup the test environment.
-     */
-    public function setUp(): void
-    {
-        parent::setUp();
-        app()['config']->set('h-rbac.rbacClass', TestAuthorizationClass::class);
-    }
-
-
-    /**
      * @param \Illuminate\Foundation\Application $app
      *
      * @return array